requirements.txt
python-telegram-bot
mercadopago
httpx
requests
anyio
certifi
httpcore
idna
sniffio
h11
charset-normalizer
urllib3
